requirements.txt
ctransformers
gradio
huggingface_hub
langchain
python-dotenv
tensorflow
torch
transformers
